requirements.txt
Flask==3.1.0
pymongo[srv]==4.10.1
flask-login==0.6.3
python-dotenv==1.0.1
certifi